refactor: remove helmut.manager.object and helmut.manager.jobs.job
