+++ pytest_splunk_addon/helmut/manager/jobs/sdk/job.py
@@ -19,16 +19,20 @@
 @since: 2011-11-23
 """
 import logging
+import time
 
 import splunklib.results as results
 
-from pytest_splunk_addon.helmut.manager.jobs.job import Job
 from pytest_splunk_addon.helmut.manager.jobs.results import Results
+from pytest_splunk_addon.helmut.exceptions.search import SearchFailure
+from pytest_splunk_addon.helmut.exceptions.wait import WaitTimedOut
 
 LOGGER = logging.getLogger("helmut")
 
 
-class SDKJobWrapper(Job):
+class SDKJobWrapper:
+    _SECONDS_BETWEEN_JOB_IS_DONE_CHECKS = 1
+
     def __init__(self, sdk_connector, sdk_job):
         """
         The constructor of the SDKJobWrapper.
@@ -39,8 +43,7 @@ class SDKJobWrapper(Job):
         @type param: splunklib.client.Job
         """
         self._raw_sdk_job = sdk_job
-
-        super(SDKJobWrapper, self).__init__(sdk_connector)
+        self._connector = sdk_connector
 
     @property
     def raw_sdk_job(self):
@@ -81,6 +84,46 @@ class SDKJobWrapper(Job):
         response = self.raw_sdk_job.results(**kwargs)
         return _build_results_from_sdk_response(response)
 
+    def wait(self, timeout=5400):
+        """
+        Waits for this search to finish.
+
+        @param timeout: The maximum time to wait in seconds. None or 0
+                        means no limit, None is default.
+        @type timeout: int
+        @return: self
+        @rtype: L{SDKJobWrapper}
+        @raise WaitTimedOut: If the search isn't done after
+                                  C{timeout} seconds.
+        """
+        LOGGER.debug("Waiting for job to finish.")
+        if timeout == 0:
+            timeout = None
+
+        start_time = time.time()
+        while not self.is_done():
+            try:
+                if self.is_failed():
+                    LOGGER.warn(
+                        "job %s failed. error message: %s"
+                        % (self.sid, self.get_messages())
+                    )
+                    break
+            except AttributeError as e:
+                LOGGER.debug(str(e))
+            _check_if_wait_has_timed_out(start_time, timeout)
+            time.sleep(self._SECONDS_BETWEEN_JOB_IS_DONE_CHECKS)
+
+        LOGGER.debug("Job %s wait is done." % self.sid)
+        return self
+
+    def check_message(self):
+        if self.get_messages():
+            message = self.get_messages()
+            for key in message:
+                if key == "error":
+                    raise SearchFailure(message[key])
+
 
 def _build_results_from_sdk_response(response):
     """
@@ -101,3 +144,14 @@ def _build_event_from_results_reader(reader):
     for field in list(reader.keys()):
         event[field] = reader[field]
     return event
+
+
+def _check_if_wait_has_timed_out(start_time, timeout):
+    if timeout is None:
+        return
+    if _wait_timed_out(start_time, timeout):
+        raise WaitTimedOut(timeout)
+
+
+def _wait_timed_out(start_time, timeout):
+    return time.time() > start_time + timeout
